multiboot2: Implement TXT slaunch support
The code makes sure that MBI entry goes first into DRTM policy, so the
payload can measure it the first thing on launch.

SLRT table is initially allocated on the heap, then an area for it is
reserved inside TXT heap by TXT code and the data is eventually copied
into its final place.  This is done to be able to start building SLRT
before TXT heap is setup.

To avoid mismatch between OS SINIT data MLE size and MLE size from
MLE header, take the mle_size from MLE header as is (without aligning
it).

Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>
Signed-off-by: Tomasz Żyjewski <tomasz.zyjewski@3mdeb.com>
Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>
Signed-off-by: Sergii Dmytruk <sergii.dmytruk@3mdeb.com>