i386/efi: Add DL stub as common DL event module
The purpose of the stub is allowing DLME to call back into GRUB to
initiate a dynamic launch.  GRUB essentially starts DLME twice: first
without and then with DRTM.  This allows DLME to properly initialize
itself using EFI services before DRTM is initiated.

To avoid duplication legacy boot is performed through the stub as well,
but GRUB calls it directly and DLME is started only once.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>
Signed-off-by: Sergii Dmytruk <sergii.dmytruk@3mdeb.com>